+++ src/main/java/com/sillim/recordit/enums/date/WeekNumber.java
@@ -1,4 +1,4 @@
-package com.sillim.recordit.schedule.domain;
+package com.sillim.recordit.enums.date;
 
 import com.fasterxml.jackson.annotation.JsonValue;
